created new suspense wrapper components for both student and instrument card lists
