+++ src/app/dashboard/search/page.tsx
@@ -7,8 +7,10 @@ import { useAppSelector } from "@/app/lib/ReduxSSR/hooks";
 import { RootState } from "@/app/lib/ReduxSSR/store";
 
 //component imports
-import CardList from "../../components/card-list/cardList";
+import CardList from "../../components/card-list/studentCardList";
 import Loading from "../../components/loading/loading";
+import StudentCardListSuspenseWrapper from "../../components/card-list/studentCardList";
+import InstrumentCardListSuspenseWrapper from "@/app/components/card-list/instrumentCardList";
 
 export default function Search() {
  
@@ -49,24 +51,9 @@ export default function Search() {
   });
   
   return (      
-    <section className="flex flex-col basis-3/4 items-center">
-      {selectOption === "Search Instrument" &&  (
-        <Suspense fallback={<Loading/>}>
-          <CardList
-            instrumentSearchResults={instrumentSearchResults}
-            selectOption={selectOption}
-          />
-        </Suspense>
-        )}
-
-      {selectOption === "Search Student" && (
-        <Suspense fallback={<Loading/>}>
-          <CardList
-            studentSearchResult={studentSearchResults}
-            selectOption={selectOption}
-          />
-        </Suspense>
-        )}        
+    <section className="flex flex-col basis-3/4 items-center">     
+      {selectOption === "Search Student" && <StudentCardListSuspenseWrapper studentSearchResult={studentSearchResults}/>}  
+      {selectOption === "Search Instrument" && <InstrumentCardListSuspenseWrapper instrumentSearchResults={instrumentSearchResults}/>}       
     </section>    
   );
 }
